Edit eslint
run: npx esLint .

 changed for  

 uses: eslint/github-action@v0
GitHub executes the src/github-action.js file. That file, in turn, calls npx @eslint/github-action, using the version number found in package.json, to execute the logic. By publishing the GitHub Action to npm and then running it with npx, we avoid the need to check dependencies into the repository.